streamRead / streamWrite: Avoid memcpy and local buffer
Build a slice that is backed by the passed buffer rather using memcpy
to move data around.

This fixes a segfault in TestReaderBZIP2 with go1.8 where a large
chunk of data is written through (*Rules).Write() ->
yr_rules_save_stream() -> streamWrite() and piped to an external bzip2
process. (This *may* have had to do with a the stack being moved due
to the large buffer allocation.)