Implement API that allows providing data for modules.
Most YARA modules use the data being scanned as their only input, but some others needs additional data (example: the "cuckoo" module). With the existing API the user didn't have a way for providing additional data to modules. This commit implements a new set of API functions that accept a ScanOptions structure, where one the fields is a map of strings to byte slices that provides additional data for modules.

As the new API is more general, the existing functions where implemented using the new ones.