New api (#3)
* Add Scan{Mem,File,Proc,FileDescriptor}WithCallback methods, ScanCallback interface

The (*rules).ScanX functions are now implemented in terms of the
corresponding (*rules).ScanXWithCallback methods.

* Implement Scanner API and split the ScanCallback interface into multiple ones.

This commit implements a new API for leveraging the new functions YARA's C API (https://github.com/VirusTotal/yara/commit/be0fc3f52f140d38a6a45db006ddc2b71d4de825). Additionally the ScanCallback interface is split into multiple ones with a single function each. This allows client to implement only the function they are interested in, instead of being forced to implement a lot of functions.

Another relevant change is the introduction of a scanContext structure that helps keeping track of buffers with module data that need to be released when the scan finishes.