Fix bug in Match.Length method.
The length returned by this method should be the actual length for the match, but it was returning the length of the (possibly truncated) copy of the matching data that YARA maintains. The maximum length of the copied data can be configured with SetMaxMatchData and it can be even zero.