Replace cbpool with cgo.Handle (or a private backport thereof)
"runtime/cgo".Handle introduced in go 1.17 is a decent solution for
the problem solved by the callback data pool. Since it's desirable to
support Go < 1.17, add a compatibility shim using a backport of
"runtime/cgo".Handle from Go 1.18's stdlib..

Thanks to Max Altgelt for the idea and the initial patch (#99).