ScanMem*: Keep buffer or iterator objects alive
After getting reports about crashes in a scanner that uses ScanMem on
mmap'd files, I have come to the conclusion that this may have to do
with the GC inteerfering with the buffer.

Adding runtime.KeepAlive should not incur much of an overhead.

(I haven't been able to reproduce the issue, though.)