feat: keep alive structs for C usage
When accessing C fields or functions, there was a potential race
condition with the garbage collector running the finalizer if
the object was no longer referenced elsewhere. This race was
(mostly) short, but arguably still existed.
Use KeepAlive() to safeguard these accesses and ensure the object
is only garbage collected (and the C part freed) once the access
has finished.